requirements.txt
asgiref==3.4.1
blis==0.7.6
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
class-registry==2.1.2
click==8.0.4
cymem==2.0.6
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
environ==1.0
filters==1.3.2
gunicorn==20.1.0
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.6
nltk==3.7
numpy==1.22.3
packaging==21.3
pathy==0.6.1
preshed==3.0.6
psycopg2==2.9.2
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2021.11.10
requests==2.27.1
six==1.16.0
smart-open==5.2.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.15
tqdm==4.63.0
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.9
wasabi==0.9.0
whitenoise==5.3.0
